+++ djangocms_blog/cms_toolbar.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 from cms.toolbar_base import CMSToolbar
 from cms.toolbar_pool import toolbar_pool
+from cms.utils.urlutils import admin_reverse
 from django.core.urlresolvers import reverse
 from django.utils.translation import override, ugettext_lazy as _
 
@@ -33,6 +34,22 @@ class BlogToolbar(CMSToolbar):
                     'admin:djangocms_blog_post_change', args=(current_post.pk,)),
                     active=True)
 
+    def add_publish_button(self):
+        """
+        Adds the publish button to the toolbar if the current post is unpublished
+        """
+        current_post = getattr(self.request, get_setting('CURRENT_POST_IDENTIFIER'), None)
+        if (self.toolbar.edit_mode and current_post and
+                not current_post.publish and
+                self.request.user.has_perm('djangocms_blog.change_post')
+            ):  # pragma: no cover  # NOQA
+            classes = ['cms-btn-action', 'blog-publish']
+            title = _('Publish {0} now').format(current_post.app_config.object_name)
+
+            url = admin_reverse('djangocms_blog_publish_article', args=(current_post.pk,))
+
+            self.toolbar.add_button(title, url=url, extra_classes=classes, side=self.toolbar.RIGHT)
+
     def post_template_populate(self):
         current_post = getattr(self.request, get_setting('CURRENT_POST_IDENTIFIER'), None)
         if current_post and self.request.user.has_perm('djangocms_blog.change_post'):  # pragma: no cover  # NOQA
@@ -52,3 +69,4 @@ class BlogToolbar(CMSToolbar):
                 menu.remove_item(pagetags)
             except ImportError:
                 pass
+            self.add_publish_button()