+++ djangocms_blog/migrations/0007_auto_20150719_0933.py
@@ -1,18 +1,23 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.db import models, migrations
-import filer.fields.image
-import django.utils.timezone
-from django.conf import settings
-import djangocms_text_ckeditor.fields
 import django.db.models.deletion
+import django.utils.timezone
+import djangocms_text_ckeditor.fields
+import filer.fields.image
+from django.conf import settings
+from django.db import migrations, models
+from djangocms_blog.models import thumbnail_model
+from filer.settings import FILER_IMAGE_MODEL
+
+ACTUAL_FILER_IMAGE_MODEL = FILER_IMAGE_MODEL or 'filer.Image'
 
 
 class Migration(migrations.Migration):
 
     dependencies = [
         ('djangocms_blog', '0006_auto_20150214_1907'),
+        migrations.swappable_dependency(ACTUAL_FILER_IMAGE_MODEL),
     ]
 
     operations = [
@@ -99,19 +104,19 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='post',
             name='main_image',
-            field=filer.fields.image.FilerImageField(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image', to='filer.Image', related_name='djangocms_blog_post_image', null=True),
+            field=filer.fields.image.FilerImageField(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image', to=ACTUAL_FILER_IMAGE_MODEL, related_name='djangocms_blog_post_image', null=True),
             preserve_default=True,
         ),
         migrations.AlterField(
             model_name='post',
             name='main_image_full',
-            field=models.ForeignKey(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image full', to='cmsplugin_filer_image.ThumbnailOption', related_name='djangocms_blog_post_full', null=True),
+            field=models.ForeignKey(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image full', to=thumbnail_model, related_name='djangocms_blog_post_full', null=True),
             preserve_default=True,
         ),
         migrations.AlterField(
             model_name='post',
             name='main_image_thumbnail',
-            field=models.ForeignKey(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image thumbnail', to='cmsplugin_filer_image.ThumbnailOption', related_name='djangocms_blog_post_thumbnail', null=True),
+            field=models.ForeignKey(on_delete=django.db.models.deletion.SET_NULL, blank=True, verbose_name='main image thumbnail', to=thumbnail_model, related_name='djangocms_blog_post_thumbnail', null=True),
             preserve_default=True,
         ),
         migrations.AlterField(